requirements.txt
# flask
flask-cors
# gunicorn
keras==2.15.0
tensorflow==2.15.0
pillow
opencv-python
# torch
# torchvision